requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==2.2.2
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-google-genai==2.1.9
google-generativeai==0.8.3
langchain-community==0.3.26
streamlit==1.37.1
pinecone>=6.0.0,<8.0.0
-e.


